light-testcase-parameters: Rename setup directory to testcase_parameters
- everything was deleted from setup/ dir except testcase_parameters.py
- so directory could be renamed
